[Lint] Blame it all on Eloraam
This fixes a ton of style issues and warnings. It also fixes Mixin javadoc authors and reasons. (or at least guesses upon what they are)

Eloraam is *usually* the author of Forge patches, but it would take forever to check the author of every single one. So I just credited her for now.